# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Booking Scheduler API - V1",
        "description": "Complete API documentation for Smart Booking Scheduler. Comprehensive system for managing services, bookings, and scheduling.",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "rdhathaliya411@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/working-hours/special-days": {
            "get": {
                "tags": [
                    "Working Hours"
                ],
                "summary": "Fetch all special day rules",
                "description": "Get list of all special day configurations with pagination, sorting, and filters.",
                "operationId": "getSpecialDays",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortOrder"
                    },
                    {
                        "$ref": "#/components/parameters/Search"
                    },
                    {
                        "$ref": "#/components/parameters/Status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Special days fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Special days fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/SpecialDay"
                                                    }
                                                },
                                                "pagination": {
                                                    "$ref": "#/components/schemas/Pagination"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Working Hours"
                ],
                "summary": "Create a special day rule",
                "description": "Add a special day rule such as full off, half day, or extra working hours.",
                "operationId": "createSpecialDay",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "type"
                                ],
                                "properties": {
                                    "date": {
                                        "description": "Special day date in Y-m-d format",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "type": {
                                        "description": "Type of special day",
                                        "type": "string",
                                        "enum": [
                                            "full_off",
                                            "half_day",
                                            "extra_hours"
                                        ],
                                        "example": "extra_hours"
                                    },
                                    "half_day_type": {
                                        "description": "Required only when type = half_day",
                                        "type": "string",
                                        "enum": [
                                            "am",
                                            "pm"
                                        ],
                                        "example": "am",
                                        "nullable": true
                                    },
                                    "start_time": {
                                        "description": "Required only when type = extra_hours",
                                        "type": "string",
                                        "example": "13:00",
                                        "nullable": true
                                    },
                                    "end_time": {
                                        "description": "Required only when type = extra_hours and must be after start_time",
                                        "type": "string",
                                        "example": "16:00",
                                        "nullable": true
                                    },
                                    "title": {
                                        "description": "Optional title for special day",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "New Year Eve",
                                        "nullable": true
                                    },
                                    "comment": {
                                        "description": "Optional comment",
                                        "type": "string",
                                        "example": "Office closes early",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Special day created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Special day created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SpecialDay"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/working-hours/special-days/{id}": {
            "delete": {
                "tags": [
                    "Working Hours"
                ],
                "summary": "Delete a special day rule",
                "description": "Deletes a special day by ID.",
                "operationId": "deleteSpecialDay",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the special day to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Special day deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Special day deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Special day not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Special day not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Pagination": {
                "title": "Pagination Meta",
                "description": "Pagination information for list responses",
                "properties": {
                    "current_page": {
                        "description": "Current page number",
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "description": "Items per page",
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "description": "Total number of items",
                        "type": "integer",
                        "example": 156
                    },
                    "last_page": {
                        "description": "Last page number",
                        "type": "integer",
                        "example": 11
                    },
                    "from": {
                        "description": "First item index on page",
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "description": "Last item index on page",
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
...
            "SpecialDay": {
                "title": "Special Day",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "title": {
                        "type": "string",
                        "example": "Christmas Eve"
                    },
                    "comment": {
                        "type": "string",
                        "example": "Extended break"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "type": {
                        "description": "Special day type",
                        "type": "string",
                        "enum": [
                            "full_off",
                            "half_day",
                            "extra_hours"
                        ],
                        "example": "extra_hours"
                    },
                    "half_day_type": {
                        "description": "Used only when type = half_day",
                        "type": "string",
                        "enum": [
                            "am",
                            "pm"
                        ],
                        "example": "am",
                        "nullable": true
                    },
                    "start_time": {
                        "description": "Required only when type = extra_hours",
                        "type": "string",
                        "example": "13:00",
                        "nullable": true
                    },
                    "end_time": {
                        "description": "Required only when type = extra_hours and must be after start_time",
                        "type": "string",
                        "example": "16:00",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "ValidationError": {
                "description": "Request validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Validation failed"
                                },
                                "errors": {
                                    "type": "object",
                                    "example": {
                                        "email": [
                                            "The email field is required."
                                        ],
                                        "password": [
                                            "Password must be at least 8 characters."
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "Unauthorized": {
                "description": "Authentication required or invalid/expired token",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "ServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Internal server error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Page number (1-indexed)",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                }
            },
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Number of items per page (max 100)",
                "required": false,
                "schema": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1,
                    "example": 15
                }
            },
            "SortBy": {
                "name": "sort_by",
                "in": "query",
                "description": "Field name to sort by",
                "required": false,
                "schema": {
                    "type": "string",
                    "example": "created_at"
                }
            },
            "SortOrder": {
                "name": "sort_order",
                "in": "query",
                "description": "Sort direction",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ],
                    "example": "desc"
                }
            },
            "Search": {
                "name": "search",
                "in": "query",
                "description": "Search keyword",
                "required": false,
                "schema": {
                    "type": "string",
                    "example": "hair cut"
                }
            },
            "Status": {
                "name": "status",
                "in": "query",
                "description": "Filter by status",
                "required": false,
                "schema": {
                    "type": "boolean"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based authentication",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}